requirements.txt
langchain-google-genai~=2.1.4
pandas==2.3.2
google-genai==1.39.1